move factored out function
